requirements.txt
Pillow==10.1.0
cairosvg==2.7.1
